# backend/docs/swagger.yaml
basePath: /
definitions:
...
  entity.CreateWorkflowReq:
    properties:
      description:
        maxLength: 1000
        type: string
      is_default:
        type: boolean
      name:
        maxLength: 255
        minLength: 2
        type: string
    required:
    - name
    type: object
  entity.CreateWorkflowStatusReq:
    properties:
      category:
        enum:
        - todo
        - in_progress
        - done
        type: string
      color:
        type: string
      is_initial:
        type: boolean
      name:
        maxLength: 100
        minLength: 1
        type: string
      position:
        type: integer
    required:
    - category
    - name
    type: object
...
  entity.UpdateWorkflowReq:
    properties:
      description:
        maxLength: 1000
        type: string
      is_default:
        type: boolean
      name:
        maxLength: 255
        minLength: 2
        type: string
    type: object
  entity.UpdateWorkflowStatusReq:
    properties:
      category:
        enum:
        - todo
        - in_progress
        - done
        type: string
      color:
        type: string
      name:
        maxLength: 100
        minLength: 1
        type: string
      position:
        type: integer
    type: object
...
  entity.Workflow:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      id:
        type: string
      is_default:
        type: boolean
      name:
        type: string
      statuses:
        items:
          $ref: '#/definitions/entity.WorkflowStatus'
        type: array
      transitions:
        items:
          $ref: '#/definitions/entity.WorkflowTransition'
        type: array
      updated_at:
        type: string
    type: object
  entity.WorkflowStatus:
    properties:
      category:
        description: todo | in_progress | done
        type: string
      color:
        type: string
      created_at:
        type: string
      id:
        type: string
      is_initial:
        type: boolean
      name:
        type: string
      position:
        type: integer
      updated_at:
        type: string
      workflow_id:
        type: string
    type: object
  entity.WorkflowTransition:
    properties:
      created_at:
        type: string
      from_status:
        $ref: '#/definitions/entity.WorkflowStatus'
      from_status_id:
        description: nil = from any status
        type: string
      id:
        type: string
      name:
        type: string
      to_status:
        $ref: '#/definitions/entity.WorkflowStatus'
      to_status_id:
        type: string
      workflow_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Project management backend API (Jira-like).
  title: JiraFlow API
  version: "1.0"
paths:
...
  /api/v1/workflows:
    get:
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/entity.Workflow'
                type: array
              limit:
                type: integer
              page:
                type: integer
              total:
                type: integer
              total_pages:
                type: integer
            type: object
      security:
      - BearerAuth: []
      summary: List workflows
      tags:
      - workflows
    post:
      consumes:
      - application/json
      parameters:
      - description: Workflow data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.CreateWorkflowReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              data:
                $ref: '#/definitions/entity.Workflow'
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              code:
                type: string
              message:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create workflow
      tags:
      - workflows
  /api/v1/workflows/{id}:
    delete:
      parameters:
      - description: Workflow ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            properties:
              code:
                type: string
              message:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete workflow
      tags:
      - workflows
    get:
      parameters:
      - description: Workflow ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/entity.Workflow'
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              code:
                type: string
              message:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get workflow with statuses and transitions
      tags:
      - workflows
    put:
      consumes:
      - application/json
      parameters:
      - description: Workflow ID
        in: path
        name: id
        required: true
        type: string
      - description: Update data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.UpdateWorkflowReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/entity.Workflow'
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              code:
                type: string
              message:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update workflow
      tags:
      - workflows
  /api/v1/workflows/{workflow_id}/statuses:
    post:
      consumes:
      - application/json
      parameters:
      - description: Workflow ID
        in: path
        name: workflow_id
        required: true
        type: string
      - description: Status data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.CreateWorkflowStatusReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              data:
                $ref: '#/definitions/entity.WorkflowStatus'
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              code:
                type: string
              message:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add status to workflow
      tags:
      - workflows
...
  /api/v1/workflows/statuses/{id}:
    delete:
      parameters:
      - description: Status ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            properties:
              code:
                type: string
              message:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete workflow status
      tags:
      - workflows
    put:
      consumes:
      - application/json
      parameters:
      - description: Status ID
        in: path
        name: id
        required: true
        type: string
      - description: Update data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.UpdateWorkflowStatusReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/entity.WorkflowStatus'
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              code:
                type: string
              message:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update workflow status
      tags:
      - workflows
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"